requirements.txt
attrs==23.2.0
certifi==2024.6.2
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
PySocks==1.7.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
wsproto==1.2.0
PyYAML~=6.0.1